Add files via upload
This is expanding on some of the math Ryan was doing. Some of the units are incorrect, and I'm going through and adjusting them as time goes on. I think I'm going to have this file plot out a possible flow model for the injector. Idk, very much under development.